Refactor mock dag into it's own header for re-use
